requirements.txt
Django==3.2
Pillow==9.1
django-ckeditor
